moved data in json files
