+++ src/commands/admin/setprefix.js
@@ -1,6 +1,6 @@
 // The two next lines can't failed because they have been already executed in index.js
 const Discord = require("discord.js");
-var data = require("../../data");
+var data = require("../../data/myserver.json");
 
 const timelimit = 360000; // 1 hour
