+++ src/commands/admin/setup.js
@@ -1,6 +1,6 @@
 // The two next lines can't failed because they have been already executed in index.js
 const Discord = require("discord.js");
-var data = require("../../data");
+var data = require("../../data/myserver.json");
 
 module.exports = {
     description: "Command fully to setup the bot",
