+++ src/data/default.json
@@ -1,0 +1,4 @@
+{
+    "prefix": "!",
+    "color": "#a1107d"
+}
